requirements.txt
paddlepaddle-gpu>=2.5.0
Pillow>=9.0.0
opencv-python>=4.7.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy==1.24.4
